+++ internal/volume/volume.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 )
 
 const (
@@ -22,6 +23,12 @@ const (
 	distDirContainer    = "/app/" + distRelativePath
 	tmpDirContainer     = "/app/" + tmpRelativePath
 	workDirContainer    = "/app"
+
+	// cacheDisabledValue is the value user can specify to disable cache
+	// The possibility to disable cache has been introduced to mitigate some
+	// issues on docker for Windows (v2.2.0.5) handing files on shared volumes.
+	// See https://github.com/lucor/fyne-cross/issues/54
+	cacheDisabledValue = "no"
 )
 
 // Volume represents the fyne-cross projec layout
@@ -49,15 +56,14 @@ func Mount(workDirHost string, cacheDirHost string) (*Volume, error) {
 		if err != nil {
 			return nil, fmt.Errorf("Cannot get the path for the system cache directory on the host %s", err)
 		}
-		cacheDirHost = filepath.Join(userCacheDir, fyneCrossPrefix)
-
+		cacheDirHost = JoinPathHost(userCacheDir, fyneCrossPrefix)
 	}
 
 	l := &Volume{
-		binDirHost:   filepath.Join(workDirHost, binRelativePath),
+		binDirHost:   JoinPathHost(workDirHost, binRelativePath),
 		cacheDirHost: cacheDirHost,
-		distDirHost:  filepath.Join(workDirHost, distRelativePath),
-		tmpDirHost:   filepath.Join(workDirHost, tmpRelativePath),
+		distDirHost:  JoinPathHost(workDirHost, distRelativePath),
+		tmpDirHost:   JoinPathHost(workDirHost, tmpRelativePath),
 		workDirHost:  workDirHost,
 	}
 
@@ -134,4 +140,21 @@ func (l *Volume) TmpDirContainer() string {
 // WorkDirContainer returns the working dir on the host
 func (l *Volume) WorkDirContainer() string {
 	return workDirContainer
+}
+
+// CacheEnabled returns true if the cache is enabled
+func (l *Volume) CacheEnabled() bool {
+	return l.CacheDirHost() != cacheDisabledValue
+}
+
+// JoinPathContainer joins any number of path elements into a single path,
+// separating them with the Container OS specific Separator.
+func JoinPathContainer(elem ...string) string {
+	return strings.Join(elem, "/")
+}
+
+// JoinPathHost joins any number of path elements into a single path,
+// separating them with the Host OS specific Separator.
+func JoinPathHost(elem ...string) string {
+	return filepath.Join(elem...)
 }