+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.5.0] - 2020-04-13
+- Add android support #37
+- Add iOS support on Darwin hosts
+- Issue cross compiling from Windows 10 #54
+- Update to golang-cross:1.13.10 image (go v1.13.10)
+- Update to fyne cli v1.2.4
+
 ## [1.4.0] - 2020-03-04
 - Add ability to package with an icon using fyne/cmd #14
 - Update to golang-cross:1.13.8 image (go v1.13.8) #46